Add Rails 7 with Tailwind
We're going to be building a basic rails application that can be used
by someone to get an idea of how a rails app works and play around
with it on their local machine.

I've decided to stick with the default SQLite database adapter. This
is because the target audience is someone who is inexperienced and
trying this for the first time. This way we do not need to burden
them with the added complexity of setting up a database.

I considered using Tailwind for the CSS as it would allow me to
rapidly bootstrap some styles, however I opted against this for the
same reason as above: Keeping it simple for a first time user.