Add RSpec for testing
The demo application is going to have some logic where it pulls listings
from an API and then presents that to the user through the user interface. It
will also let them like the listings.

I'm opting to do this with a test driven development (TDD) approach. TDD is a
highly encouraged practice as it makes is think about what we want our code to
do ahead of writing it, and ensure it works as expected. It will be good for
these to be in place for the end user to be able to see while they're learning.

I've opted for RSpec as this is a popular and well supported testing framework.