Add rubocop
```
RuboCop is a Ruby static code analyzer (a.k.a. linter) and code formatter. Out of the box it will enforce many of the guidelines outlined in the community Ruby Style Guide.
```

I've opted to add rubocop to help check that the code meets the
[Ruby Style Guide](https://rubystyle.guide/) and to help demonstrate this
to the end user of this app.

In addition I've included modules for Rails (this framework) and RSpec (our testing framework).