Add a sensible rubocop configuration
Although rubocop is great, it's default configuration can be a little overkill.
I've added the configuration I use on other projects to make it a little easier to work with.

This config is opinionated and there could be any number of valid arguments for changing any of the settings. If people decide to make changes in the future that's totally fine, this has just been added as a starting point.