Add ArticlesService
We need to process a hosted JSON file and process it into usable ruby 
objects.

In previous commits I prepared for this by:
- Adding a Article model
- Adding the Faraday gem

Now I am adding a service class called ArticlesService. It's job is to 
fetch the JSON file and then convert each article into a Article model.

This app is small and this could've been done in the controller, but I 
have opted for a service class because:

- They are easier to test than controllers
- We want to have 'skinny' controllers 
  https://dev.to/kputra/rails-skinny-controller-skinny-model-5f2k#phase-3

In future commits we'd want to:

- catch and handle any exceptions raised by Faraday
- define how we return those errors from the service to the controller 
  so that it could display a friendly message to the user.