Add ArticleLikes controller
I am working on a feature that will allow a user to add a like to an article. These are tracked in ArticleLike records.

In preperation for that feature we need a controller to handle saving that record.

I've added an ArticleLikes controller that will create a like using the article_id, and will redirect the user back to the articles page.

In a future commit we could integrate Turbo so that the user remains on the same page.