Add likes count to articles page
In the previous commit I added the likes count to the `Article` class. Now I display that on the articles page.

I stopped using instance_doubles for `Article`s in the articles feature spec since they are looking up data now. I didn't want to stub this out in a feature spec as then I wouldn't really be testing that everything fits together and works as expected.