Add Article model
I am preparing to fetch a list of articles from an API and then display them on a page.

Ahead of this work I have created an Article object. I'll create an Article for each item fetched from the hosted JSON file.

I've opted for this approach so that I can test the article, and have an easy way to test any future validation that may be done.

I'll add more fields to it later as they become required by other parts of the application.