Add ArticlesController
Not that the ArticlesService and Article 'model' are in place I have
created the ArticlesController. Its job is to use the ArticlesService
to fetch articles from the hosted json file, and then assign them to
the view.

I've included some basic styling for the ArticlesController#index page
which I'll expand on in future commits.

In future commits we'd want to handle what happens when:
- No articles are returned, i,e, display a friendly message
- An error is returned by the ArticlesService