Add article_likes table & model
I'm adding a feature where a user can like a post displayed on the
`/articles` page.

I needed a place to store the likes and have opted to create an
article_likes page. I've named it `article_likes` to highlight which
other model it relates to.

When designing this table I considered two approaches:

1. Have a single record associated with an article id, and update it
   when new likes are added
2. Create a new record for each like

Each have benefits and downsides.

1: would have faster queries as the data grows, however we would need
   to mitigate for possible race conditions.
2: will have exponentially more records, however we won't need to worry
   about race conditions. However 2 will allow us to track more info in
   the future, like the IP address of the like.

I opted for option 2 for its potential future benefits, and included an
index to help keep the look-ups fast.