Add likes count to Article model
I'm going to display the number of likes that an article has alongside
it on the `/articles` page.

The `ArticlesController#index` view already has access to the `Article`
so I've added a `likes` method to an Article so that it will be really
easy to get the number of likes in the view.